[level 1] Title: 정수 제곱근 판별, Time: 0.09 ms, Memory: 80.4 MB -BaekjoonHub

+++ 프로그래머스/1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -1,7 +1,7 @@
 class Solution {
     public long solution(long n) {
         long answer = 0;
-        double result = Math.sqrt((double) n);
+        double result = Math.sqrt(n);
         
         if (isInteger(result)) {
             result = result + 1;
